Made it so that it can now get average speed
I also modified the speed_threshold to 10, because it is influencing the average running speed. When I had it set to 8.5, my average running speed was around 8.5 8.6, but by inspecting the graph it is clearly between 9 and 11. I also changed it so that when calculating the speed it does a rolling average instead of cutting out data.